feat: implement ActionIntentPromptBuilder and update PromptBuilder for purpose-based prompt generation
