+++ docs/architecture/prompt-system.md
@@ -10,12 +10,12 @@ The architecture explicitly avoids persona as one giant prompt only; see `docs/c
 
 | Purpose | Description | V1 status |
 | --- | --- | --- |
-| `action_intent` | Builds the existing full persona prompt for choosing an action intent and optional visible chat text. | Active |
-| `social_interpretation` | Reserved for interpreting tone, ambiguity, sarcasm, passive aggression, and plausible motive. | Reserved |
-| `background_reflection` | Reserved for relationship memory, emotional residue, and pending-intention consolidation. | Reserved |
-| `spectator_recap` | Reserved for narrator-facing recap generation. | Reserved |
+| `action_intent` | Builds the existing full persona prompt for choosing an action intent and optional visible chat text. | Active via `ActionIntentPromptBuilder` |
+| `social_interpretation` | Reserved for interpreting tone, ambiguity, sarcasm, passive aggression, and plausible motive. | Reserved; no builder yet |
+| `background_reflection` | Reserved for relationship memory, emotional residue, and pending-intention consolidation. | Reserved; no builder yet |
+| `spectator_recap` | Reserved for narrator-facing recap generation. | Reserved; no builder yet |
 
-Reserved values exist to make field gating explicit. They must not be wired into runtime LLM calls until their prompt builders and tests exist.
+Reserved values exist to make field gating explicit. They must not be wired into runtime LLM calls or passed to the current `PromptBuilder.build()` until their prompt builders and tests exist.
 
 ## Field-Purpose Matrix
 
@@ -33,7 +33,9 @@ Reserved values exist to make field gating explicit. They must not be wired into
 
 Every `BuiltPrompt` carries its `purpose`. Providers can receive one prompt object without needing extra parameters, and observability can derive the call type from the same source that built the prompt.
 
-The default purpose is `action_intent` to preserve the existing runtime behavior and tests.
+There is no implicit default at the prompt-construction boundary. Call sites must pass a purpose explicitly so new LLM surfaces cannot accidentally inherit the action prompt.
+
+`PromptBuilder` is a dispatcher, not a universal prompt template. Each supported purpose must have a dedicated builder. Today only `ActionIntentPromptBuilder` exists.
 
 ## callType Alignment
 
@@ -50,7 +52,7 @@ The default purpose is `action_intent` to preserve the existing runtime behavior
 
 1. Add or confirm the value in `PromptPurpose`.
 2. Define the exact allowed persona fields for that purpose.
-3. Add prompt-building logic that excludes unrelated fields.
+3. Add a dedicated prompt builder that excludes unrelated fields and uses the correct output contract.
 4. Add a `purposeToCallType` mapping case.
 5. Add tests for included fields, excluded fields, output contract, and usage call type.
 6. Update this document with the new purpose and invariants.
@@ -61,3 +63,4 @@ The default purpose is `action_intent` to preserve the existing runtime behavior
 - Raw chain-of-thought must not be requested, stored, or exposed.
 - Agents must never see another agent's private reasoning, private motive summaries, or spectator narration.
 - `action_intent` prompt output must remain stable unless the action prompt is intentionally changed and tested.
+- Reserved purposes must fail closed until their dedicated prompt surfaces exist.
